Added a number of new widgets and utility functions for use in the Review Board UI rewrite.
The main highlights include an auto-size text area, inline editor, modal
dialog box, and tooltips. There's also a utility function for
introducing a delay in an animation sequence and one for positioning an
element to the side of another element intelligently based on the sizes
of both.

Reviewed at http://reviews.review-board.org/r/658/


git-svn-id: http://svn.navi.cx/misc/trunk/djblets@11918 9a66fdd2-3ab9-0310-9df6-91dd0304fdbc