Prevent the default event when clicking to activate an inlineEditor.
If an inlineEditor's target element was inside of a container responding
to clicks (such as an <a href="...">), then clicking the element would
activate the inlineEditor and then trigger the click action of the container.
This would lead to page reloads and other problems, depending on the
situation. We now prevent the default action from happening at all.

Fixes bug #1221.