Fix auto-size text area issues with Firefox, Google Chrome and Safari.
There were some issues with the auto-size text areas that were pretty bad.

The first and most majorist one was that with Google Chrome, every key
press would grow the text area, making it unusable. This was due to a
calculation bug with WebKit, which was worked around thanks to a patch by
Cory McWilliams.

The second issue is that on Firefox, hitting backspace would cause the
text area to resize too small by about one line, or sometimes more depending
on content. This seems to be due to switching over to needing pre-wrap on
newer Firefoxes instead of -moz-pre-wrap. We now default to pre-wrap.

We had previously disabled auto-size text areas whenever running with Safari.
Now we allow Safari to run, and instead disable auto-sizing for versions of
WebKit older than 531.9.

Tested on IE, Firefox, Chrome and Safari.