Fix image cropping with the new Django storage support.
Python Image Library's Image class doesn't set Image.format when
creating a new cropped image. Since we now save to a temp file at first, and
that doesn't contain a file extension, PIL has no idea what format to save
the image as, causing the cropping to fail. We now pass in the format we used
when parsing the filename.

To help catch problems in the future, we also now log any and all exceptions
during cropping or thumbnailing.

Fixes bug #1396