Fix screenshot cropping in djblets templatetags.
The crop_image templatetag was broken in a few subtle ways. The first big
breakage was that the scoping got changed so it was using an undefined variable
for the name returned, which would cause it to print <img src="" ...>. Second,
we were sometimes passing in bad format keys, which made PIL barf. Since we know
that these are going to be simple (and generated by our code), we might as well
always save as PNG. Third, there was a broken format string in the logging code,
which prevented it from printing the exception to the log :/

This fixes Review Board bug 1490