Change the edit columns button to use a pencil icon.
The edit columns button in the datagrid had a "..." label on them, but that
wasn't very obvious as to its purpose. Given that a pencil icon is pretty
synonymous with "edit" in many webapps (Review Board included), it seems
suitable for an "edit columns" icon as well. This change switches to that.

Fixes bug #910