Add support for easy pagination with the web API.
This introduces a new WebAPIPaginated class that returns a response with
built-in pagination. It takes a queryset and some optional values for the
keys in the payload, default number of entries returned in a response, and
maximum number of entries that can be requested in a single response.

The class handles 'start' and 'max-results' parameters to the URL, which
specify the first item's index (0-based) and the maximum number of items
that will be returned.

The response payload contains the results as a list, the total number of
objects (used to calculate how many requests would be needed), and
URLs for previous and next queries.

Reviewed at http://reviews.reviewboard.org/r/1460/