Add new __version__ and __version_info__ to djblets.
The __version__ variable contains the human-readable version string
(which is the results of get_package_version()).

The __version_info__ variable contains all the info from VERSION except
for the release flag.

These are still not standard variables, but are semi-standard at least,
and would likely be tried first before VERSION or get_package_version().