Fix rendering issues with columns in the datagrids.
Adding columns to a datagrid would cause odd rendering issues. This was due
to not including the 'last' flag (indicating the last column, in order to
position the Edit Columns column properly) in the cache key. We now add
the key, which allows the datagrid to render properly when adding or
removing columns.