Add support for Selenium unit testing.
This adds support for easily adding Selenium unit tests to a test suite.
It contains a version of the Python Selenium driver for talking to a
Selenium RC server, and some functionality for starting/stopping a live
dev server.

Test suites can instantiate djblets.testing.SeleniumUnitTest and access
self.selenium in order to talk to the Selenium server. The test suite will
handle starting/stopping the dev server and the browser session. By default,
this attempts to reuse the dev server and browser session (which other
Selenium test suite implementations do not appear to do), though this can be
customized per-suite.