Store a cell's URL and CSS class as part of the cache key.
We were storing too little in the cache key for cells. If two cells had the
same rendered content, but had different CSS classes or URLs, then they would
still be treated as identical cells. This would lead to cells linking to the
wrong URLs or displaying the wrong CSS.

We now calculate the URL and CSS class before building the cache key, and
include that data as part of the key. Only the template rendering is done
inside the cache entry creation code path.