Change the format for hrefs.
Instead of 'child_hrefs', we now have 'links', which map link names
to information about the link. The information is a dictionary containing
'href' and 'method' (the HTTP method).

'links' contains a special 'self' name that maps to the actual
location for that resource. It also contains entries for the operations
that can be performed on the resource ('create', 'update', 'delete').

This replaces both 'child_hrefs' and 'href'.

All links are now absolute URIs, instead of relative to the API base or
to the resource.