Return the resource tree's list of URI templates for resources in the root.
This is a nice little shortcut for clients that want to be very RESTful
and not hard-code URIs, but at the same time don't want to walk the
resource tree to find a leaf resource they want. They can now query the
list of URI templates from the root resource, plug in the values they
want, and reference that URL directly.

This is a pattern I haven't seen done anywhere else, but should be very
beneficial to clients accessing the Review Board API, as the resource
tree can get deep.