Resource names should now use '_' instead of '-'.
Instead of resource names containing '-' characters, they're now supposed
to contain '-' characters. These will be normalized when building URIs.
The reason for this is to make it easy for JavaScript clients to access
the contents in the links array by referencing the variables directly
instead of accessing by a key string.