Copy the headers dictionary in WebAPIResponseError.
We were defaulting the headers parameter to the WebAPIResponseError
constructor to None and then creating a new dictionary if it was None. The
reason for this is that we modify the dictionary and don't want to risk
that going into the static default dictionary for the parameter (as this
would impact future instances). However, the right solution is just to
copy the dictionary.