Make it easier to develop resources that act as singletons.
Previously, creating a resource that returned a single instance of an object
(instead of a list of objects or an object by ID) was difficult, and broke
such things as child resources. Now, an object can set itself as a singleton
(singleton = True) and everything will work as expected. get() will be called,
but not get_list().