Fix the selenium test skip exception when the Selenium RC is not installed.
When the Selenium RC isn't installed, the first time SeleniumUnitTest.setUp is
called it will raise SkipTest, but subsequent calls wouldn't go through this
code path because cls.selenium was still set to non-None. Any other calls to
cls.selenium would raise an HTTP connection error, but that got reported as an
error instead of properly skipping the test.