Extension enabling/disabling API, with UI
I've created an Extensions resource for the Djblets API.  Authorized users can
get the entire list of extensions from a GET on /api/extensions/, or info on a
particular extension through a GET on /api/extensions/[extension id]/.
Enabling and disabling extensions is accomplished through a PUT on
/api/extensions/[extension id]/ with an "enabled" boolean parameter.

The UI has been updated to use this API through AJAX queries.  Errors during
enabling/disabling should be displayed in a modal dialog box (see screenshot).

Some of the CSS, and a few of the images have been copied from the RB project.
Not sure if we want that duplication.  Just a heads up.

Another note - I've added a new field to the RegisteredExtensions table.  This
means, if you're following along with me, you'll need to perform an evolve on
your DB.

Reviewed at http://reviews.reviewboard.org/r/1618/