Add a CounterField for atomically updating counters.
This provides a new CounterField, which can be used for atomically
incrementing or decrementing an integer. It's intended for use as
a substitute to potentially expensive or numerous SQL queries to retrieve
counts of objects.

CounterField can take an optional initializer function to populate the
field the first time it's ever accessed (such as after creation or
database evolution). After that, the value will be treated as an integer.

CounterField can update an individual instance's field, or can batch-update
the field on several objects at once.

Reviewed at http://reviews.reviewboard.org/r/1835/