Fix CounterField for new models.
Any new model with a CounterField was being saved immediately upon creation,
which would cause lots of problems, particularly when adding a new object in
the Admin UI.

Now, new objects are initialized to 0. They don't use the initializer
(since the initializer will likely access fields on the model that can't
yet be accessed). We may want to extend this longer-term to use better
defaults or to try an initializer and fall back on 0.

This also gets rid of the CounterProxy class, which shouldn't have slipped in.