Don't error out in the gravatar templatetag if the user is anonymous.
The {% gravatar %} templatetag would end up raising an exception (and emailing
the server admin) if called with an anonymous user, because the AnonymousUser
class doesn't have an "email" field. This would ideally never happen in real
life, except it is because the review board user page is busted. Trivial fix.