Don't error out if a JSONField is initialized with blank in kwargs.
Our JSONField.__init__ method unconditionally initialized the TextField base
with blank=True, but also passed through kwargs un-modified. This meant that if
a user also added the blank kwarg, it would error out. This is fixed by popping
blank out of the kwargs first, and then passing it back in.

Fixes bug 2015.