Fix a storage problem with large cache data.
Large cache data was encoded incorrectly, causing anything stored to become
unloadable, meaning it'd have to be recomputed. This was due to a string being
stored for all but the last chunk, instead of a string wrapped in a list.
This affects any data that, after being picked and zlib-compressed, is over
1MB in size.

There are now unit tests that check for this.

Fixes bug #1660