Fix a date range calculation for the log viewer.
The "This Month" view in the server log viewer was broken due to it attempting
to use the beginning of the date range calculated, which was 0, instead of 01.
Since we can assume the start of the month will continue to be 1 for the
foreseeable future, we can just hardcode this and be safe.

Fixes bug #2218