Add a custom TestCase class and an @add_fixtures decorator for tests.
Tests that are on a Djblets TestCase subclass can augment or replace the
fixtures they use through the @add_fixtures decorator. By default, this will
add to the list of fixtures, but if replace=True is passed, it can replace
the existing list.

This allows test cases to be more specific about their fixture needs, which
can speed up tests by not including the union of all required fixtures in
TestCase.fixtures.