Consolidate the jquery script includes into two templates.
We had a lot of places both in Djblets and in Review Board that
hard-coded paths to the jquery scripts. This made it hard to perform
upgrades, relocate to an off-site script, etc.

Now we have two templates responsible for including the scripts. These
can be easily updated, and projects using Djblets can override the
templates to provide a different path or version if necessary.

This is the precursor to some work for updating the versions of jquery
and jquery-ui we use.