Added support for displaying a field with no label.
In some case, label is not needed for fields of form. For example, if we
want to display a formset, it is boring to display the same label for
every forms. To disable label, the form class need a field
"fields_no_label", and set it to True.

Reviewed at http://http://reviews.reviewboard.org/r/2671/