Re-land localized timezone support.
Django 1.4 introduces timezone awareness and localization. This updates
Djblets to take advantage of this when available by creating
timezone-aware datetime obnjects.

DateTimeColumn and DateTimeSinceColumn are updated to accept a new
parameter "timezone", which converts the datetime object into that
timezone. A TimeZoneField form field is created to make implementing a
timezone ChoiceField simple. Lastly, timezones are created as UTC and
are checked to be aware/naive for backwards compatibility.

As part of this, we now require Django 1.4.

Written by Dave Druska, with changes by Christian Hammond to remove the
compatibility work for older versions of Django.

Reviewed at http://reviews.reviewboard.org/r/2721/