Fix parsing of tokens in a blocktag with Django 1.4.
We were failing to pass a tuple to parser.parse() when specifying the
end tags for a block tag. Somehow this never bit us before, but a
seemingly unrelated change in Django 1.4 triggered this broken behavior,
making all custom blocktags break.

This led to some false-positives in other fixes, where it seemed that
removing custom tags starting with "if" solved things. In reality, it
didn't solve anything. It just masked the problem.

This change should be compatible with all versions of Django.