Switch Djblets to using Django's static media setup.
Django introduced support for a static files framework in 1.3, which is
becoming more commonplace amongst other Django apps. It separates things
into "static media" (media files bundled with the project/apps,
designated by STATIC_* settings) and uploaded media (which the old
MEDIA_* settings are now used for).

This updates Djblets to work with this scheme. It requires that the
application have some basic support for static media set up.

This change may not be compatible with older Review Board releases, but
those releases will end up hard-coding support for 0.6.x.