Localized timezone support (Djblets)
Supports the USE_TZ feature in Django dev by creating UTC, timezone aware
datetime objects.

DateTimeColumn and DateTimeSinceColumn are updated to accept a new parameter
"timezone", which converts the datetime object into that timezone. A
TimeZoneField form field is created to make implementing a timezone ChoiceField
simple. Lastly, timezones are created as UTC and are checked to be aware/naive
for backwards compatibility.