Switch fully to new-style static support.
The new Djblets release will be fully moved on top of Django's static
media support, with a heavy preference toward using Pipeline. Since that
gives us file versioning, there's really no need to require MEDIA_SERIAL
anymore, so long as we properly use {% static %} when we want to
reference a file.

So now all MEDIA_SERIAL usage is gone. The functions that generate the
media serials remain, as does the context processor for getting it,
but they're entirely opt-in now, and deprecated. They may end up raising
deprecation warnings later.