Add a calculated dirty state and a dirtyStateChanged signal to inlineEditor.
inlineEditor now calculates the dirty state on various events and stores
it. This makes it cheaper to look up, and also provides a central point
to notify consumers of when the dirty state changes.

It also fixes the Escape key in certain browsers, by way of how the key
event handling changed.

Reviewed at http://reviews.reviewboard.org/r/3000/