Lazily import User and AnonymousUser in djblets_utils.
rbssh in Review Board makes use of humanize_list in djblets_utils, but
the recent introduction of the User and AnonymousUser imports broke it,
as it then required Django settings to be set up.

This moves the import into the one function that needs it, in order to
prevent this kind of problem.

Fixes bug #2686