Fix link generation for extension web API resources
This change adds the Web API resources defined in an extension's
`resources` attribute to the link tree. Now each resource will be added
to the 'links' dictionary of the extension resource, and in the list of
extensions. These links will only appear when the extension is enabled.

Also fixed is improper generation of the resource URLs. The URLs were
generated with the wrong attribute, causing them to use underscores
instead of dashes.

Reviewed at http://reviews.reviewboard.org/r/3082/