Conditionalize ez_setup
This makes life simpler for packagers, since most packaging tools
will ensure that the appropriate packages are available on the
system. Additionally, most packaging tools prefer not to rely on
network access (since it can mean that identical rebuilds cannot
be performed).

Built locally with 'python setup.py build' and also by applying this patch and
performing an RPM build in a network-free environment.

This is essentially the same patch as
http://reviews.reviewboard.org/r/3186/ (which was for Review Board
itself)

Reviewed at http://reviews.reviewboard.org/r/3252/