Stop polluting the global WebAPIResource.allowed_mimetypes list.
We were adding items to self.allowed_mimetypes in WebAPIResource, but
this was the class-defined list, and not an instance-defined list. This
meant that every single WebAPIResource indicated in this field that it
accepted every known mimetype. We now make a copy of the list early on
for the instance to modify.