Add facilities for reloading gravatar images in retina resolution.
This adds a jQuery function which, when applied to an <img class="gravatar">
will parse the URL and reload it at the appropriate resolution for
high-resolution devices. This requires some javascript on the reviewboard side
as well, to call it in the relevant places.

Tested in conjunction with a corresponding reviewboard change.
Reviewed at http://reviews.reviewboard.org/r/3482/