Handle setting a dictionary in siteconfig's cache_backend.
siteconfig's cache_backend was expecting the stored data to be a string
representing an old-style cache backend URI. However, when first
creating a new site now, settings.CACHES will be stored in siteconfig,
which breaks our bad assumptions.

Now when loading from siteconfig, if cache_backend is a dictionary with
a 'default' entry, we just pull from it and set it.

This allows either an old-style string or a new-style CACHES info
dictionary to be stored and accessed.