Add a normalize_cache_backend function.
This provides a function for taking either a new-style CACHES dictionary
or an old-style CACHE_BACKEND and returning a valid CACHES dictionary.
It's basically just logic moved out of siteconfig and into a new
djblets.util.cache.

Now that we have this file, I'll later begin work on transitioning our
other caching functions.