Fix redisplaying the inline editor a second time.
We had a bug where the second time an inline editor was opened, the
buttons would appear on the next row and the field would be short. This
was due to us hiding the editor before setting the CSS display to
inline. Since we hid first, jQuery saved the "old" display as the
default, "block," and restored this on show later.

Now we set the display type before calling hide, in order to ensure it's
restored properly.