Conditionally delete "info" and "registration" attributes from an extension.
When an extension is known about but not yet handled, it doesn't yet
have "info" and "registration" attributes on the class. However, doing a
reload of the extensions (which will happen the first time we go to load
extensions, or when an extension is enabled/disabled in some thread)
will cause us to try to delete these attributes. If they don't exist, we
fail.

Now we're checking if these exist before trying to remove them.