Fix the z-indexes for modalboxes.
Modalboxes had a z-index with a very large number, which placed it on
top of everything. It used 11000 and 11001, which were kind of silly.
Move these down to 99 and 100, making it easier to work with.

By a wild coincidence, this places the dialog at exactly the right
stacking order for Review Board.