Provide a Admin UI setting to change Log Levels
Allow Logging level to be setup in Admin UI

Currently, the logging level can only be setup only in settings_local.py which
makes it difficult to change the value. Provide a dropdown in the Admin->Logging
Screens that shows the various logging levels.

Added default value for logging_level in siteconfig.py as well as mapped the
field shown in the UI to the internal variable in siteconfiguration tables.

This patch goes along with the Reviewboard change at #3852
(http://reviews.reviewboard.org/r/3852/)

Testing done:
Verified in fresh setup (devserver) where it defaults to the DEBUG level.  Since
this only gets stored in siteconfig_siteconfiguration table, shouldn't need any
migration changes. Changed the level manually in the UI and verified that DEBUG
messages dont appear anymore.

Also the Filter in Admin->Server Log shows only the LogLevel selected and
anything above it. For e.g. if the logging is set to ERROR, the filters in
Server Log only show Error, Critical and ALL.

Addresses bug 1859.
Reviewed at http://reviews.reviewboard.org/r/3851/