Install an extension on-the-fly from a remote source.
Added a method to djblet's extension manager that installs an extension,
given a remote URL with the extension egg and package name.

This modification is required as a pre-requisite, for the upcoming
extension browser module to ReviewBoard.

There could be possible errors such as (but not limited to): The URL
being a 404, incorrect package_name, incompatible egg or the package
being already present. These would be thrown as exceptions and have to
be handled by the caller appropriately.

Reviewed at http://reviews.reviewboard.org/r/3973/