Fix configuration of the STATIC_URL in siteconfig.
Our siteconfig code was properly turning the site_static_url setting
into settings.STATIC_URL, but it's too late at this point. Django has
already created the StaticFilesStorage and populated it with the default
settings.STATIC_URL.

Fortunately, this all happens before that URL is ever used (since this
is before any templates are rendered). So we can just grab the
staticfiles_storage instance and replace the URL.